📝 (UserService.java): Remove o método save e sua lógica de tratamento de exceção ConflictException 🐛 (CreateUserService.java): Adiciona verificação de existência de usuário por e-mail antes de criar um novo usuário
