+++ user/src/main/java/com/javanauta/user/business/UserService.java
@@ -5,7 +5,6 @@ import com.javanauta.user.business.dto.request.UserRequestDTO;
 import com.javanauta.user.business.dto.response.UserResponseDTO;
 import com.javanauta.user.core.utils.JwtUtil;
 import com.javanauta.user.infrastructure.entities.User;
-import com.javanauta.user.infrastructure.exceptions.ConflictException;
 import com.javanauta.user.infrastructure.exceptions.ResourceNotFoundException;
 import com.javanauta.user.infrastructure.repositories.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -28,21 +27,6 @@ public class UserService {
         return userConverter.toResponse(user);
     }
 
-    public UserResponseDTO save(UserRequestDTO dto) {
-        try {
-            this.existsByEmail(dto.getEmail());
-
-            User user = userConverter.toEntity(dto);
-            user.setPassword(passwordEncoder.encode(dto.getPassword()));
-
-            User inserted = this.userRepository.save(user);
-            return this.userConverter.toResponse(inserted);
-        } catch (ConflictException e) {
-            throw new ConflictException(e.getMessage(), e.getCause());
-
-        }
-    }
-
     public UserResponseDTO updateCurrentUserData(UserRequestDTO dto, String token) {
         String email = jwtUtil.extractUserEmailFromToken(token.split(" ")[1]);
         User userEntity = this.userRepository.findByEmail(email).orElseThrow(
@@ -58,16 +42,5 @@ public class UserService {
 
     public void deleteByEmail(String email) {
         this.userRepository.deleteByEmail(email);
-    }
-
-    private void existsByEmail(String email) {
-        try {
-            boolean exists = userRepository.existsByEmail(email);
-            if (exists) {
-                throw new ConflictException("User with email " + email + " already exists");
-            }
-        } catch (ConflictException e) {
-            throw new ConflictException("User with email " + email + " already exists", e.getCause());
-        }
     }
 }
